drivers: Introduce brand new kernel gaming mode
How to trigger gaming mode? Just open a game that is supported in games list
How to exit? Kill the game from recents, or simply back to homescreen

What does this gaming mode do?
- It limits big cluster maximum frequency to 2,0GHz, and little cluster
  to values matching GPU frequencies as below:
+ 338MHz: 455MHz
+ 385MHz and above: 1053MHz
- As for the cluster freq limits, it overcomes heating issue while playing
  heavy games, as well as saves battery juice

Big thanks to [kerneltoast] for the following commits on his wahoo kernel
https://github.com/kerneltoast/android_kernel_google_wahoo/commit/5ac1e81d3d9e77803cdd6714e70c4463e386ac4c
https://github.com/kerneltoast/android_kernel_google_wahoo/commit/e13e2c4554124e94fbee5bb7fc1582d301e41015

Gaming control's idea was based on these

Signed-off-by: Diep Quynh <remilia.1505@gmail.com>